smal tweaks

+++ src/camera.h
@@ -96,7 +96,7 @@ s32 get_viewport_count(void);
 s32 get_current_viewport(void);
 void func_80066230(Gfx **dlist, MatrixS **arg1);
 f32 get_distance_to_active_camera(f32 xPos, f32 yPos, f32 zPos);
-void func_800663DC(s32 xPos, s32 yPos, s32 zPos, s32 arg3, s32 arg4, s32 arg5);
+void func_800663DC(f32 xPos, f32 yPos, f32 zPos, s32 arg3, s32 arg4, s32 arg5);
 void write_to_object_render_stack(s32 arg0, f32 xPos, f32 yPos, f32 zPos, s16 arg4, s16 arg5, s16 arg6);
 s8 check_if_showing_cutscene_camera(void);
 s32 set_active_viewports_and_max(s32 arg0);
